Move CPU shaders to the crates/shaders
