+++ Cargo.toml
@@ -55,6 +55,7 @@ workspace = true
 
 [dependencies]
 vello_encoding = { workspace = true }
+vello_shaders = { workspace = true }
 bytemuck = { workspace = true }
 skrifa = { workspace = true }
 peniko = { workspace = true }
@@ -71,6 +72,7 @@ clippy.semicolon_if_nothing_returned = "warn"
 
 [workspace.dependencies]
 vello_encoding = { version = "0.1.0", path = "crates/encoding" }
+vello_shaders = { version = "0.1.0", path = "crates/shaders" }
 bytemuck = { version = "1.15.0", features = ["derive"] }
 skrifa = "0.19.0"
 peniko = "0.1.0"
