+++ crates/shaders/src/cpu/coarse.rs
@@ -6,11 +6,9 @@ use vello_encoding::{
     DRAW_INFO_FLAGS_FILL_RULE_BIT,
 };
 
-use crate::cpu_dispatch::CpuBinding;
-
 use super::{
-    CMD_BEGIN_CLIP, CMD_COLOR, CMD_END, CMD_END_CLIP, CMD_FILL, CMD_IMAGE, CMD_JUMP, CMD_LIN_GRAD,
-    CMD_RAD_GRAD, CMD_SOLID, CMD_SWEEP_GRAD, PTCL_INITIAL_ALLOC,
+    CpuBinding, CMD_BEGIN_CLIP, CMD_COLOR, CMD_END, CMD_END_CLIP, CMD_FILL, CMD_IMAGE, CMD_JUMP,
+    CMD_LIN_GRAD, CMD_RAD_GRAD, CMD_SOLID, CMD_SWEEP_GRAD, PTCL_INITIAL_ALLOC,
 };
 
 const N_TILE_X: usize = 16;
