+++ crates/shaders/src/cpu/euler.rs
