+++ crates/shaders/src/cpu/fine.rs
@@ -3,9 +3,7 @@
 
 use vello_encoding::{ConfigUniform, PathSegment, Tile};
 
-use crate::cpu_dispatch::CpuTexture;
-
-use super::{CMD_COLOR, CMD_END, CMD_FILL, CMD_JUMP, CMD_SOLID, PTCL_INITIAL_ALLOC};
+use super::{CpuTexture, CMD_COLOR, CMD_END, CMD_FILL, CMD_JUMP, CMD_SOLID, PTCL_INITIAL_ALLOC};
 
 // These should also move into a common area
 const TILE_WIDTH: usize = 16;
