+++ crates/shaders/src/cpu/path_tiling_setup.rs
@@ -3,7 +3,7 @@
 
 use vello_encoding::{BumpAllocators, IndirectCount};
 
-use crate::cpu_dispatch::CpuBinding;
+use super::CpuBinding;
 
 const WG_SIZE: usize = 256;
