+++ crates/shaders/src/cpu/util.rs
