+++ crates/shaders/src/lib.rs
@@ -5,11 +5,12 @@ mod types;
 
 #[cfg(feature = "compile")]
 pub mod compile;
-
-pub use types::{BindType, BindingInfo, WorkgroupBufferInfo};
+#[cfg(feature = "cpu")]
+pub mod cpu;
 
 #[cfg(feature = "msl")]
 pub use types::msl;
+pub use types::{BindType, BindingInfo, WorkgroupBufferInfo};
 
 use std::borrow::Cow;
