+++ src/lib.rs
@@ -81,10 +81,6 @@
 //!
 //! See the [`examples/`](https://github.com/linebender/vello/tree/main/examples) folder to see how that code integrates with frameworks like winit and bevy.
 
-#[cfg(feature = "wgpu")]
-mod cpu_dispatch;
-#[cfg(feature = "wgpu")]
-mod cpu_shader;
 mod recording;
 mod render;
 mod scene;
